Safari-609.1.11.1/LayoutTests/http/tests/security/xss-DENIED-synchronous-form-expected.txt

CONSOLE MESSAGE: line 9: SecurityError: Blocked a frame with origin "http://localhost:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
Waiting for iframe...
Done.
